Post fix

+++ app/controllers/books_controller.rb
@@ -1,17 +1,22 @@
 class BooksController < ApplicationController
+	before_action :authenticate_user!
+
 	def index
 		@books = Book.all
-		@book = Book.new
+		@new_book = Book.new
+		@user = User.find(current_user.id)
 	end
 
 	def show
 		@book = Book.find(params[:id])
+		@new_book = Book.new
+		@user = User.find(current_user.id)
 	end
 
 	def create
 		book = Book.new(book_params)
 		if book.save
-			flash[:success] = 'successfully'
+			flash.now[:success] = 'successfully'
 			redirect_to book_path(book.id)
 		else
 			@books = Book.all
@@ -25,6 +30,7 @@ class BooksController < ApplicationController
 
 	def update
 		@book = Book.find(params[:id])
+		raise StandardError if current_user.id != @book.user.id
 		if @book.update(book_params)
 			redirect_to book_path(@book.id)
 		else
@@ -34,11 +40,12 @@ class BooksController < ApplicationController
 
 	def destroy
 		book = Book.find(params[:id])
+		raise StandardError if current_user.id != book.user.id
 		if book.destroy
-			flash[:success] = '削除しました'
+			flash.now[:success] = '削除しました'
 			redirect_to books_path
 		else
-			flash[:danger] = '削除できませんでした'
+			flash.now[:danger] = '削除できませんでした'
 			redirect_to books_path
 		end
 	end
@@ -46,6 +53,6 @@ class BooksController < ApplicationController
 	private
 
 	def book_params
-		params.require(:book).permit(:title, :body)
+		params.require(:book).permit(:title, :body).merge(user_id: current_user.id)
 	end
 end
